updated with grid and new buttons layout for  layers
