--- package/Editor/LayerEditor/ContextMenuActions.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 05aba7df292542f4688a5b7d13d00810
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
