--- package/Editor/LayerEditor/LayerEditor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d49909f6e8a3eac49b56d4bd0ee29f20
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
