--- package/Editor/LayerEditor/LayerEditorRightPanel.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1396980cd97bda64681edc51d0f1d29a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
